Added bosses, and enemies per level
Added a system to determine which enemies should be present in which
level. This way,  you will only encounter stronger monsters in higher
levels, if so described in the textfile.
Also added bosses. One boss will be chosen to be the final boss and the
player will be able to encounter him in the last room of the dungeon!